backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.7.0
langchain==0.0.336
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3